requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-google-genai
pillow
pymupdf
python-dotenv
openai
langchain-openai
huggingface_hub
langchain-huggingface
# torch
# transformers
# accelerate
# sentencepiece
# protobuf
# scipy